+++ src/tools/pulse/listMetricDefinitions/listPulseMetricDefinitionsFromDefinitionIds.test.ts
@@ -1,7 +1,7 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
-import type { PulseMetricDefinition } from '../../sdks/tableau/types/pulse.js';
-import { server } from '../../server.js';
+import type { PulseMetricDefinition } from '../../../sdks/tableau/types/pulse.js';
+import { server } from '../../../server.js';
 import { listPulseMetricDefinitionsFromDefinitionIdsTool } from './listPulseMetricDefinitionsFromDefinitionIds.js';
 
 // Mock server.server.sendLoggingMessage since the transport won't be connected.
@@ -32,7 +32,7 @@ const mocks = vi.hoisted(() => ({
   mockListPulseMetricDefinitionsFromMetricDefinitionIds: vi.fn(),
 }));
 
-vi.mock('../../restApiInstance.js', () => ({
+vi.mock('../../../restApiInstance.js', () => ({
   getNewRestApiInstanceAsync: vi.fn().mockResolvedValue({
     pulseMethods: {
       listPulseMetricDefinitionsFromMetricDefinitionIds:
