+++ src/tools/pulse/listMetricSubscriptions/listPulseMetricSubscriptions.ts
@@ -1,9 +1,9 @@
 import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 import { Ok } from 'ts-results-es';
 
-import { getConfig } from '../../config.js';
-import { getNewRestApiInstanceAsync } from '../../restApiInstance.js';
-import { Tool } from '../tool.js';
+import { getConfig } from '../../../config.js';
+import { getNewRestApiInstanceAsync } from '../../../restApiInstance.js';
+import { Tool } from '../../tool.js';
 
 export const listPulseMetricSubscriptionsTool = new Tool({
   name: 'list-pulse-metric-subscriptions',
