+++ src/tools/pulse/listMetrics/listPulseMetricsFromMetricDefinitionId.ts
@@ -2,9 +2,9 @@ import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 import { Ok } from 'ts-results-es';
 import { z } from 'zod';
 
-import { getConfig } from '../../config.js';
-import { getNewRestApiInstanceAsync } from '../../restApiInstance.js';
-import { Tool } from '../tool.js';
+import { getConfig } from '../../../config.js';
+import { getNewRestApiInstanceAsync } from '../../../restApiInstance.js';
+import { Tool } from '../../tool.js';
 
 export const listPulseMetricsFromMetricDefinitionIdTool = new Tool({
   name: 'list-pulse-metrics-from-metric-definition-id',
