+++ src/tools/pulse/listMetrics/listPulseMetricsFromMetricIds.ts
@@ -2,9 +2,9 @@ import { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 import { Ok } from 'ts-results-es';
 import { z } from 'zod';
 
-import { getConfig } from '../../config.js';
-import { getNewRestApiInstanceAsync } from '../../restApiInstance.js';
-import { Tool } from '../tool.js';
+import { getConfig } from '../../../config.js';
+import { getNewRestApiInstanceAsync } from '../../../restApiInstance.js';
+import { Tool } from '../../tool.js';
 
 export const listPulseMetricsFromMetricIdsTool = new Tool({
   name: 'list-pulse-metrics-from-metric-ids',
